Moved strategy file from python/bot/fuzzers to python/fuzzing so it can be accessed through appengine for MAB cron job.
