+++ src/appengine/handlers/cron/fuzz_strategy_selection.py
@@ -21,10 +21,10 @@ combined strategies. In the upload_bandit_weights function, we can change
 metric to be for edges, crash, features, or units. Currently based on new
 edges."""
 
-from bot.fuzzers import strategy
 from datastore import data_types
 from datastore import ndb
 from datastore import ndb_utils
+from fuzzing import strategy
 from google_cloud_utils import big_query
 from handlers import base_handler
 from libs import handler
