+++ src/python/bot/fuzzers/engine_common.py
@@ -25,8 +25,8 @@ import time
 
 from base import utils
 from bot.fuzzers import options
-from bot.fuzzers import strategy
 from bot.fuzzers import utils as fuzzer_utils
+from fuzzing import strategy
 from metrics import fuzzer_stats
 from metrics import logs
 from system import archive
