+++ src/python/tests/core/bot/fuzzers/libFuzzer/libfuzzer_strategy_selection_test.py
@@ -15,11 +15,11 @@
 
 import unittest
 
-from bot.fuzzers import strategy
 from bot.fuzzers.libFuzzer import strategy_selection
 from bot.tasks import fuzz_task
 from datastore import data_types
 from datastore import ndb
+from fuzzing import strategy
 from system import environment
 from tests.test_libs import helpers as test_helpers
 from tests.test_libs import test_utils
